src/requirements.txt
azure.core
azure-ai-documentIntelligence
streamlit
azure-storage-blob
python-dotenv
altair==4